requirements.txt
pyautogui
openpyxl
pdfminer.six
pypdf


